Adds import-address sample
